Move the failure class to the report package
