+++ src/main/java/io/vertx/ext/unit/report/Failure.java
@@ -1,4 +1,4 @@
-package io.vertx.ext.unit;
+package io.vertx.ext.unit.report;
 
 import io.vertx.codegen.annotations.CacheReturn;
 import io.vertx.codegen.annotations.GenIgnore;
