+++ src/main/java/io/vertx/ext/unit/report/TestResult.java
@@ -2,7 +2,6 @@ package io.vertx.ext.unit.report;
 
 import io.vertx.codegen.annotations.CacheReturn;
 import io.vertx.codegen.annotations.VertxGen;
-import io.vertx.ext.unit.Failure;
 
 /**
  * The result of a test.
