+++ src/main/java/io/vertx/ext/unit/report/impl/EventBusReporter.java
@@ -3,7 +3,7 @@ package io.vertx.ext.unit.report.impl;
 import io.vertx.core.Vertx;
 import io.vertx.core.eventbus.MessageProducer;
 import io.vertx.core.json.JsonObject;
-import io.vertx.ext.unit.Failure;
+import io.vertx.ext.unit.report.Failure;
 import io.vertx.ext.unit.report.TestResult;
 import io.vertx.ext.unit.impl.FailureImpl;
 import io.vertx.ext.unit.report.Reporter;
